fix build - think was a circ ref
