+++ packages/data-types/src/types/v1/date.ts
@@ -7,7 +7,7 @@ import { GraphQLType, TypeESMapping } from '../../interfaces';
 
 export default class DateType extends BaseType {
     toESMapping(): TypeESMapping {
-        this.validateESMapping();
+        this._validateESMapping();
         let format: string|undefined;
 
         if (this.config.format && (
