+++ packages/data-types/src/types/v1/geo-json.ts
@@ -9,7 +9,7 @@ import { GraphQLType, TypeESMapping } from '../../interfaces';
 
 export default class GeoJSON extends BaseType {
     toESMapping(clientMetaData: ClientMetadata): TypeESMapping {
-        this.validateESMapping();
+        this._validateESMapping();
         // we need to used deprecated quadtree and strategy because CONTAINS is
         // not yet supported in 6.X or 7.X as of now
         return {
