+++ packages/data-types/src/types/v1/geo.ts
@@ -5,7 +5,7 @@ import { GraphQLType, TypeESMapping } from '../../interfaces';
 // TODO: This type is deprecated, not sure how to properly indicate it.
 export default class GeoType extends BaseType {
     toESMapping(): TypeESMapping {
-        this.validateESMapping();
+        this._validateESMapping();
         return {
             mapping: {
                 [this.field]: this.config.indexed === false ? {
