+++ packages/data-types/src/types/v1/object.ts
@@ -4,7 +4,7 @@ import { GraphQLType, TypeESMapping } from '../../interfaces';
 
 export default class ObjectType extends BaseType {
     toESMapping(): TypeESMapping {
-        this.validateESMapping();
+        this._validateESMapping();
 
         const type = this.config.array ? 'nested' : 'object';
         const typeConfig: ESTypeMapping = { type };
